2007-01-21  Eloy Duran <eloy.de.enige@gmail.com>
  * pre-install.rb,
    framework/tool/gen_bridge_doc/lib/cocoa_ref.rb,
    framework/tool/gen_bridge_doc.rb:
  - The RDoc generator now uses the framework in the build dir ;
  - Workarounds for bug, that was fixed in the last rev, removed ;
  - Check added for the individual html and ri dirs before trying to install them.

+++ src/pre-install.rb
@@ -80,11 +80,11 @@ end
 if File.exist?('framework/bridge-doc')
   # Frameworks HTML documentation 
   dstdir = "#{install_root}#{@config['documentation']}/RubyCocoa/Frameworks"
-  command "cp -R 'framework/bridge-doc/html' '#{dstdir}'"
+  command "cp -R 'framework/bridge-doc/html' '#{dstdir}'" if File.exist?('framework/bridge-doc/html')
   # Frameworks RI documentation
   basedstdir = "#{Config::CONFIG['datadir']}/ri/#{Config::CONFIG['ruby_version']}/site"
   unless File.exist?(basedstdir)
     command "mkdir -p '#{basedstdir}'"
   end
-  command "cp -R 'framework/bridge-doc/ri/OSX' '#{basedstdir}'"
+  command "cp -R 'framework/bridge-doc/ri/OSX' '#{basedstdir}'" if File.exist?('framework/bridge-doc/ri/OSX')
 end
